Adding RRef as return value for builtin operators
See #23110 for RRef design details. This commit only implements
RRef as return value for builtin operators, and RRef will communicate
between a user and the owner. More specifically, a RRef is first
created on the `dist.remote` caller, which is a user of the RRef.
Then the RRef user sends and notification to the owner to report
the fork to the owner, and the owner uses a shared_ptr to keep
the RRef alive. When the user RRef is destructed on the caller,
another notification will be sent to the owner, and the owner
can then drop it's RRef as well.

ghstack-source-id: caa0ee94304b7ddb3e2c8e6c34a51e41c58540d8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25169